Move bookings-related definitions to backend
Created a submodule to house all bookings-related stuff.
