+++ common/src/lib.rs
@@ -129,30 +129,6 @@ pub struct ReceiverConfiguration {
 }
 
 #[derive(Serialize, Deserialize, PartialEq, Debug, Clone)]
-pub struct Booking {
-    pub start_time: DateTime<Utc>,
-    pub end_time: DateTime<Utc>,
-    pub telescope_name: String,
-    pub user_name: String,
-}
-
-impl Booking {
-    pub fn overlaps(&self, other: &Booking) -> bool {
-        self.end_time >= other.start_time && self.start_time <= other.end_time
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq)]
-pub enum AddBookingError {
-    ServiceUnavailable,
-    Conflict,
-    // NotFuture - booking is entirely(?) in the past
-    // NonPositiveDuration - booking ends before it starts
-}
-
-pub type AddBookingResult = Result<u64, AddBookingError>;
-
-#[derive(Serialize, Deserialize, PartialEq, Debug, Clone)]
 pub struct Measurement {
     pub amps: Vec<f64>,
     pub freqs: Vec<f64>,
